api/requirements.txt
bs4
django-cors-headers
django-papertrail
django-extensions
dj_database_url
graphene
graphene_django
geohash2
gpxpy
jsonfield
newrelic
oauth2client
pykml
psycopg2-binary
shortuuid
sendwithus
zappa
-e git://github.com/joram/stravalib.git#egg=stravalib
-e git://github.com/wckd/python-yr.git#egg=python-yr
